+++ sdk/nodejs/index.ts
@@ -22,6 +22,7 @@ export * from "./loadBalancer";
 export * from "./loadBalancerMonitor";
 export * from "./loadBalancerPool";
 export * from "./logpushJob";
+export * from "./originCaCertificate";
 export * from "./pageRule";
 export * from "./provider";
 export * from "./rateLimit";